added homepage, display page, organized pages
