+++ src/DisplayContent.js
@@ -1,0 +1,17 @@
+import './App.css';
+import UploadContent from './UploadContent';
+import Home from './Home';
+
+function DisplayContent() {
+  return (
+    <div className="App">
+      <header className="App-body">
+        <div>
+          <p> display </p>
+        </div>
+      </header>
+    </div>
+  );
+}
+
+export default DisplayContent;
